+++ public/components/ltr_model/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { LtrModelListingWithRoute as LtrModelListing } from './views/ltr_model_listing';
+export { LtrModelView } from './views/ltr_model_view';
+export { LtrModelTable } from './components/ltr_model_table';
+export { LtrModelDefinition, formatDefinition } from './components/ltr_model_definition';
+export { LtrFeatureTable } from './components/ltr_feature_table';
+export { useLtrModelList } from './hooks/use_ltr_model_list';
+export { useLtrModelView } from './hooks/use_ltr_model_view';
+export * from './types';